@@ -1,2 +1,2 @@
-from .mflike import MFLike
+from .mflike import MFLike, TestMFLike
 from .theoryforge_MFLike import TheoryForge_MFLike
